% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowed_functions.R
\name{call_function}
\alias{call_function}
\title{Functions borrowed from \code{\link{tarchetypes}}}
\usage{
call_function(name, args)
}
\description{
Functions borrowed from \code{\link{tarchetypes}}
}
\keyword{borrowed}
\keyword{internal}
