% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_vue_csvs.R
\name{csv_read_in}
\alias{csv_read_in}
\title{Read-in function}
\usage{
csv_read_in(csv_batch)
}
\arguments{
\item{csv_batch}{a batch of vdat CSV files}
}
\description{
Read-in function
}
\keyword{internal}
