% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_batcher}
\alias{file_batcher}
\title{Batch a vector of files}
\usage{
file_batcher(files, batch_size)
}
\arguments{
\item{files}{character vector of files}
}
\description{
Batch a vector of files
}
\keyword{internal}
