% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_vdats.R
\name{tar_vdat_dir}
\alias{tar_vdat_dir}
\title{Convert VDAT to CSV and track}
\usage{
tar_vdat_dir(vdat_batch, csv_outdir)
}
\arguments{
\item{vdat_batch}{a batch of VDAT file paths}

\item{csv_outdir}{file path to the output directory}
}
\value{
character vector of directory file paths holding VDAT-exported CSVs
}
\description{
Convert VDAT to CSV and track
}
\keyword{internal}
\keyword{vdat}
