% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_vdats.R
\name{vdat_read_in}
\alias{vdat_read_in}
\title{Read VDAT data into a merged list by data type}
\usage{
vdat_read_in(vdat_dir)
}
\arguments{
\item{vdat_dir}{directory of VDAT-exported CSVs}
}
\value{
list of data types merged across receiver VDAT files in \code{vdat_dir}
}
\description{
Read VDAT data into a merged list by data type
}
\keyword{internal}
\keyword{vdat}
